requirements.txt
asgiref==3.4.1
attrs==21.4.0
black==21.12b0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
distlib==0.3.4
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==11.3.0
filelock==3.4.2
flake8==4.0.1
identify==2.4.3
idna==3.3
importlib-metadata==4.10.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.16.0
psycopg2==2.9.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.13.0
zipp==3.6.0
